requirements.txt
# Core Framework
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-CORS==5.0.0
Werkzeug==3.1.8

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

# Authentication & Security
email-validator==2.0.0
python-dotenv==1.0.0

# Web Scraping & Automation
selenium==4.12.0
webdriver-manager==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# Async Tasks & Queue
celery==5.3.1
redis==5.0.0

# Security & Encryption
flask-limiter==3.5.0
cryptography==42.0.5
Flask-Mail==0.10.0

# Google APIs (Gmail OAuth2)
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0

# AI & LLM (Claude API)
anthropic==0.30.0

# Scheduler (Descarga mensual de fallos)
APScheduler==3.10.4

# Server & Production
gunicorn==21.2.0

# Payment (Mercado Pago via requests, sin SDK separado)

# Logging & Monitoring
python-json-logger==2.0.7

# Testing (opcional para desarrollo)
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1

# Code Quality (opcional para desarrollo)
mypy==1.4.1
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Utilities
python-dateutil==2.8.2
